:sparkles: feat: Construindo tela de cadastro para usuário do tipo admin
